MULTI-CHUNK FILE ENCRYPTION/DECRYPTION NOW WORKS!!! :tada: :tada: :tada: :rocket: :rocket: :rocket:
...because DecryptReaderToWriter() now sets plainFile to be the file
we are decrypting to if `nil` is passed in, or, if non-nil is passed
in, it writes subsequent decryptions to that same file, and returns
that file reference so it can be passed back in the next time around!
If plainFile == nil and we are not decrypting a file, plainFile is set
to os.Stdout; great for chat scenarios!